Added react router to app, refactored App.js contents into course-finder page
